requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
anthropic>=0.40.0
python-dotenv>=1.0.0

# ── Watchers ──────────────────────────────────────────────────────────────────
watchdog>=4.0.0          # filesystem_watcher.py
requests>=2.32.0         # whatsapp_watcher.py

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── Monitoring ────────────────────────────────────────────────────────────────
psutil>=6.0.0            # watchdog/watchdog.py + monitoring/system_health.py

# ── Cloud Sync (optional) ─────────────────────────────────────────────────────
# boto3>=1.34.0          # AWS S3 sync — uncomment to enable
# google-cloud-storage>=2.17.0  # GCS sync — uncomment to enable

# ── Browser MCP (optional) ────────────────────────────────────────────────────
# playwright>=1.44.0     # Uncomment + run: playwright install chromium

# ── Google Calendar MCP (optional) ───────────────────────────────────────────
# google-api-python-client>=2.120.0
# google-auth-httplib2>=0.2.0
# google-auth-oauthlib>=1.2.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.14.0
httpx>=0.27.0            # FastAPI test client
